+++ sources/org/telegram/ui/Components/TextPaintImageReceiverSpan.java
@@ -66,10 +66,10 @@ public class TextPaintImageReceiverSpan extends ReplacementSpan {
     public void draw(Canvas canvas, CharSequence charSequence, int i, int i2, float f, int i3, int i4, int i5, Paint paint) {
         canvas.save();
         if (this.alignTop) {
-            this.imageReceiver.setImageCoords((int) f, i3 - 1, this.width, this.height);
+            this.imageReceiver.setImageCoords((float) ((int) f), (float) (i3 - 1), (float) this.width, (float) this.height);
         } else {
             int i6 = this.height;
-            this.imageReceiver.setImageCoords((int) f, i3 + ((((i5 - AndroidUtilities.dp(4.0f)) - i3) - i6) / 2), this.width, i6);
+            this.imageReceiver.setImageCoords((float) ((int) f), (float) (i3 + ((((i5 - AndroidUtilities.dp(4.0f)) - i3) - i6) / 2)), (float) this.width, (float) i6);
         }
         this.imageReceiver.draw(canvas);
         canvas.restore();
